@@ -1,9 +1,14 @@
 use std::fmt::{self, Display, Formatter};
 use std::path::PathBuf;
 
+use chrono::{DateTime, Utc};
+use clap::builder::ValueParser;
+use clap::{ArgAction, Args, ColorChoice, Parser, Subcommand, ValueEnum};
 use semver::Version;
 
-use clap::{ArgAction, Args, Parser, Subcommand, ValueEnum};
+/// The character typically used to separate path components
+/// in environment variables.
+const ENV_PATH_SEP: char = if cfg!(windows) { ';' } else { ':' };
 
 /// The Typst compiler.
 #[derive(Debug, Clone, Parser)]
@@ -13,10 +18,21 @@ pub struct CliArguments {
     #[command(subcommand)]
     pub command: Command,
 
-    /// Sets the level of logging verbosity:
-    /// -v = warning & error, -vv = info, -vvv = debug, -vvvv = trace
-    #[clap(short, long, action = ArgAction::Count)]
-    pub verbosity: u8,
+    /// Set when to use color.
+    /// auto = use color if a capable terminal is detected
+    #[clap(
+        long,
+        value_name = "WHEN",
+        require_equals = true,
+        num_args = 0..=1,
+        default_value = "auto",
+        default_missing_value = "always",
+    )]
+    pub color: ColorChoice,
+
+    /// Path to a custom CA certificate to use when making network requests.
+    #[clap(long = "cert", env = "TYPST_CERT")]
+    pub cert: Option<PathBuf>,
 }
 
 /// What to do.
@@ -30,6 +46,9 @@ pub enum Command {
     /// Watches an input file and recompiles on changes
     #[command(visible_alias = "w")]
     Watch(CompileCommand),
+
+    /// Initializes a new project from a template
+    Init(InitCommand),
 
     /// Processes an input file to extract provided metadata
     Query(QueryCommand),
@@ -49,14 +68,20 @@ pub struct CompileCommand {
     #[clap(flatten)]
     pub common: SharedArgs,
 
-    /// Path to output file (PDF, PNG, or SVG)
-    pub output: Option<PathBuf>,
+    /// Path to output file (PDF, PNG, or SVG), use `-` to write output to stdout
+    #[clap(required_if_eq("input", "-"), value_parser = ValueParser::new(output_value_parser))]
+    pub output: Option<Output>,
+
+    /// Output a Makefile rule describing the current compilation
+    #[clap(long = "make-deps", value_name = "PATH")]
+    pub make_deps: Option<PathBuf>,
 
     /// The format of the output file, inferred from the extension by default
     #[arg(long = "format", short = 'f')]
     pub format: Option<OutputFormat>,
 
-    /// Opens the output file using the default viewer after compilation
+    /// Opens the output file using the default viewer after compilation.
+    /// Ignored if output is stdout
     #[arg(long = "open")]
     pub open: Option<Option<String>>,
 
@@ -64,9 +89,28 @@ pub struct CompileCommand {
     #[arg(long = "ppi", default_value_t = 144.0)]
     pub ppi: f32,
 
-    /// Produces a flamegraph of the compilation process
-    #[arg(long = "flamegraph", value_name = "OUTPUT_SVG")]
-    pub flamegraph: Option<Option<PathBuf>>,
+    /// Produces performance timings of the compilation process (experimental)
+    ///
+    /// The resulting JSON file can be loaded into a tracing tool such as
+    /// https://ui.perfetto.dev. It does not contain any sensitive information
+    /// apart from file names and line numbers.
+    #[arg(long = "timings", value_name = "OUTPUT_JSON")]
+    pub timings: Option<Option<PathBuf>>,
+}
+
+/// Initializes a new project from a template
+#[derive(Debug, Clone, Parser)]
+pub struct InitCommand {
+    /// The template to use, e.g. `@preview/charged-ieee`
+    ///
+    /// You can specify the version by appending e.g. `:0.1.0`. If no version is
+    /// specified, Typst will default to the latest version.
+    ///
+    /// Supports both local and published templates.
+    pub template: String,
+
+    /// The project directory, defaults to the template's name
+    pub dir: Option<String>,
 }
 
 /// Processes an input file to extract provided metadata
@@ -102,21 +146,42 @@ pub enum SerializationFormat {
 /// Common arguments of compile, watch, and query.
 #[derive(Debug, Clone, Args)]
 pub struct SharedArgs {
-    /// Path to input Typst file
-    pub input: PathBuf,
+    /// Path to input Typst file, use `-` to read input from stdin
+    #[clap(value_parser = input_value_parser)]
+    pub input: Input,
 
     /// Configures the project root (for absolute paths)
     #[clap(long = "root", env = "TYPST_ROOT", value_name = "DIR")]
     pub root: Option<PathBuf>,
+
+    /// Add a string key-value pair visible through `sys.inputs`
+    #[clap(
+        long = "input",
+        value_name = "key=value",
+        action = ArgAction::Append,
+        value_parser = ValueParser::new(parse_input_pair),
+    )]
+    pub inputs: Vec<(String, String)>,
 
     /// Adds additional directories to search for fonts
     #[clap(
         long = "font-path",
         env = "TYPST_FONT_PATHS",
         value_name = "DIR",
-        action = ArgAction::Append,
+        value_delimiter = ENV_PATH_SEP,
     )]
     pub font_paths: Vec<PathBuf>,
+
+    /// The document's creation date formatted as a UNIX timestamp.
+    ///
+    /// For more information, see <https://reproducible-builds.org/specs/source-date-epoch/>.
+    #[clap(
+        long = "creation-timestamp",
+        env = "SOURCE_DATE_EPOCH",
+        value_name = "UNIX_TIMESTAMP",
+        value_parser = parse_source_date_epoch,
+    )]
+    pub creation_timestamp: Option<DateTime<Utc>>,
 
     /// The format to emit diagnostics in
     #[clap(
@@ -127,6 +192,81 @@ pub struct SharedArgs {
     pub diagnostic_format: DiagnosticFormat,
 }
 
+/// Parses a UNIX timestamp according to <https://reproducible-builds.org/specs/source-date-epoch/>
+fn parse_source_date_epoch(raw: &str) -> Result<DateTime<Utc>, String> {
+    let timestamp: i64 = raw
+        .parse()
+        .map_err(|err| format!("timestamp must be decimal integer ({err})"))?;
+    DateTime::from_timestamp(timestamp, 0)
+        .ok_or_else(|| "timestamp out of range".to_string())
+}
+
+/// An input that is either stdin or a real path.
+#[derive(Debug, Clone)]
+pub enum Input {
+    /// Stdin, represented by `-`.
+    Stdin,
+    /// A non-empty path.
+    Path(PathBuf),
+}
+
+/// An output that is either stdout or a real path.
+#[derive(Debug, Clone)]
+pub enum Output {
+    /// Stdout, represented by `-`.
+    Stdout,
+    /// A non-empty path.
+    Path(PathBuf),
+}
+
+impl Display for Output {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        match self {
+            Output::Stdout => f.pad("stdout"),
+            Output::Path(path) => path.display().fmt(f),
+        }
+    }
+}
+
+/// The clap value parser used by `SharedArgs.input`
+fn input_value_parser(value: &str) -> Result<Input, clap::error::Error> {
+    if value.is_empty() {
+        Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
+    } else if value == "-" {
+        Ok(Input::Stdin)
+    } else {
+        Ok(Input::Path(value.into()))
+    }
+}
+
+/// The clap value parser used by `CompileCommand.output`
+fn output_value_parser(value: &str) -> Result<Output, clap::error::Error> {
+    // Empty value also handled by clap for `Option<Output>`
+    if value.is_empty() {
+        Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
+    } else if value == "-" {
+        Ok(Output::Stdout)
+    } else {
+        Ok(Output::Path(value.into()))
+    }
+}
+
+/// Parses key/value pairs split by the first equal sign.
+///
+/// This function will return an error if the argument contains no equals sign
+/// or contains the key (before the equals sign) is empty.
+fn parse_input_pair(raw: &str) -> Result<(String, String), String> {
+    let (key, val) = raw
+        .split_once('=')
+        .ok_or("input must be a key and a value separated by an equal sign")?;
+    let key = key.trim().to_owned();
+    if key.is_empty() {
+        return Err("the key was missing or empty".to_owned());
+    }
+    let val = val.trim().to_owned();
+    Ok((key, val))
+}
+
 /// Lists all discovered fonts in system and custom font paths
 #[derive(Debug, Clone, Parser)]
 pub struct FontsCommand {
@@ -135,7 +275,7 @@ pub struct FontsCommand {
         long = "font-path",
         env = "TYPST_FONT_PATHS",
         value_name = "DIR",
-        action = ArgAction::Append,
+        value_delimiter = ENV_PATH_SEP,
     )]
     pub font_paths: Vec<PathBuf>,
 
@@ -160,6 +300,7 @@ impl Display for DiagnosticFormat {
     }
 }
 
+/// Update the CLI using a pre-compiled binary from a Typst GitHub release.
 #[derive(Debug, Clone, Parser)]
 pub struct UpdateCommand {
     /// Which version to update to (defaults to latest)
