@@ -3,11 +3,17 @@ use std::io::{self, Write};
 use std::path::{Path, PathBuf};
 
 use codespan_reporting::term::{self, termcolor};
+use ecow::eco_format;
 use termcolor::WriteColor;
-use typst::diag::{PackageError, PackageResult};
-use typst::syntax::PackageSpec;
+use typst::diag::{bail, PackageError, PackageResult, StrResult};
+use typst::syntax::package::{
+    PackageInfo, PackageSpec, PackageVersion, VersionlessPackageSpec,
+};
 
-use super::color_stream;
+use crate::download::{download, download_with_progress};
+use crate::terminal;
+
+const HOST: &str = "https://packages.typst.org";
 
 /// Make a package available in the on-disk cache.
 pub fn prepare_package(spec: &PackageSpec) -> PackageResult<PathBuf> {
@@ -23,18 +29,50 @@ pub fn prepare_package(spec: &PackageSpec) -> PackageResult<PathBuf> {
 
     if let Some(cache_dir) = dirs::cache_dir() {
         let dir = cache_dir.join(&subdir);
-
-        // Download from network if it doesn't exist yet.
-        if spec.namespace == "preview" && !dir.exists() {
-            download_package(spec, &dir)?;
-        }
-
         if dir.exists() {
             return Ok(dir);
+        }
+
+        // Download from network if it doesn't exist yet.
+        if spec.namespace == "preview" {
+            download_package(spec, &dir)?;
+            if dir.exists() {
+                return Ok(dir);
+            }
         }
     }
 
     Err(PackageError::NotFound(spec.clone()))
+}
+
+/// Try to determine the latest version of a package.
+pub fn determine_latest_version(
+    spec: &VersionlessPackageSpec,
+) -> StrResult<PackageVersion> {
+    if spec.namespace == "preview" {
+        // For `@preview`, download the package index and find the latest
+        // version.
+        download_index()?
+            .iter()
+            .filter(|package| package.name == spec.name)
+            .map(|package| package.version)
+            .max()
+            .ok_or_else(|| eco_format!("failed to find package {spec}"))
+    } else {
+        // For other namespaces, search locally. We only search in the data
+        // directory and not the cache directory, because the latter is not
+        // intended for storage of local packages.
+        let subdir = format!("typst/packages/{}/{}", spec.namespace, spec.name);
+        dirs::data_dir()
+            .into_iter()
+            .flat_map(|dir| std::fs::read_dir(dir.join(&subdir)).ok())
+            .flatten()
+            .filter_map(|entry| entry.ok())
+            .map(|entry| entry.path())
+            .filter_map(|path| path.file_name()?.to_string_lossy().parse().ok())
+            .max()
+            .ok_or_else(|| eco_format!("please specify the desired version"))
+    }
 }
 
 /// Download a package over the network.
@@ -43,35 +81,47 @@ fn download_package(spec: &PackageSpec, package_dir: &Path) -> PackageResult<()>
     // fetching.
     assert_eq!(spec.namespace, "preview");
 
-    let url = format!(
-        "https://packages.typst.org/preview/{}-{}.tar.gz",
-        spec.name, spec.version
-    );
+    let url = format!("{HOST}/preview/{}-{}.tar.gz", spec.name, spec.version);
 
     print_downloading(spec).unwrap();
-    let reader = match ureq::get(&url).call() {
-        Ok(response) => response.into_reader(),
+
+    let data = match download_with_progress(&url) {
+        Ok(data) => data,
         Err(ureq::Error::Status(404, _)) => {
             return Err(PackageError::NotFound(spec.clone()))
         }
-        Err(_) => return Err(PackageError::NetworkFailed),
+        Err(err) => return Err(PackageError::NetworkFailed(Some(eco_format!("{err}")))),
     };
 
-    let decompressed = flate2::read::GzDecoder::new(reader);
-    tar::Archive::new(decompressed).unpack(package_dir).map_err(|_| {
+    let decompressed = flate2::read::GzDecoder::new(data.as_slice());
+    tar::Archive::new(decompressed).unpack(package_dir).map_err(|err| {
         fs::remove_dir_all(package_dir).ok();
-        PackageError::MalformedArchive
+        PackageError::MalformedArchive(Some(eco_format!("{err}")))
     })
+}
+
+/// Download the `@preview` package index.
+fn download_index() -> StrResult<Vec<PackageInfo>> {
+    let url = format!("{HOST}/preview/index.json");
+    match download(&url) {
+        Ok(response) => response
+            .into_json()
+            .map_err(|err| eco_format!("failed to parse package index: {err}")),
+        Err(ureq::Error::Status(404, _)) => {
+            bail!("failed to fetch package index (not found)")
+        }
+        Err(err) => bail!("failed to fetch package index ({err})"),
+    }
 }
 
 /// Print that a package downloading is happening.
 fn print_downloading(spec: &PackageSpec) -> io::Result<()> {
-    let mut w = color_stream();
     let styles = term::Styles::default();
 
-    w.set_color(&styles.header_help)?;
-    write!(w, "downloading")?;
+    let mut out = terminal::out();
+    out.set_color(&styles.header_help)?;
+    write!(out, "downloading")?;
 
-    w.reset()?;
-    writeln!(w, " {spec}")
+    out.reset()?;
+    writeln!(out, " {spec}")
 }
