@@ -1,16 +1,17 @@
-use std::env;
-use std::fs;
 use std::io::{Cursor, Read, Write};
 use std::path::PathBuf;
+use std::{env, fs};
 
+use ecow::eco_format;
 use semver::Version;
 use serde::Deserialize;
 use tempfile::NamedTempFile;
-use typst::{diag::bail, diag::StrResult, eval::eco_format};
+use typst::diag::{bail, StrResult};
 use xz2::bufread::XzDecoder;
 use zip::ZipArchive;
 
 use crate::args::UpdateCommand;
+use crate::download::{download, download_with_progress};
 
 const TYPST_GITHUB_ORG: &str = "typst";
 const TYPST_REPO: &str = "typst";
@@ -20,7 +21,7 @@ const TYPST_REPO: &str = "typst";
 /// Fetches a target release or the latest release (if no version was specified)
 /// from GitHub, unpacks it and self replaces the current binary with the
 /// pre-compiled asset from the downloaded release.
-pub fn update(command: UpdateCommand) -> StrResult<()> {
+pub fn update(command: &UpdateCommand) -> StrResult<()> {
     if let Some(ref version) = command.version {
         let current_tag = env!("CARGO_PKG_VERSION").parse().unwrap();
 
@@ -50,17 +51,17 @@ pub fn update(command: UpdateCommand) -> StrResult<()> {
 
         return self_replace::self_replace(&backup_path)
             .and_then(|_| fs::remove_file(&backup_path))
-            .map_err(|err| eco_format!("failed to revert to backup: {err}"));
+            .map_err(|err| eco_format!("failed to revert to backup ({err})"));
     }
 
     let current_exe = env::current_exe().map_err(|err| {
-        eco_format!("failed to locate path of the running executable: {err}")
+        eco_format!("failed to locate path of the running executable ({err})")
     })?;
 
     fs::copy(current_exe, &backup_path)
-        .map_err(|err| eco_format!("failed to create backup: {err}"))?;
+        .map_err(|err| eco_format!("failed to create backup ({err})"))?;
 
-    let release = Release::from_tag(command.version)?;
+    let release = Release::from_tag(command.version.as_ref())?;
     if !update_needed(&release)? && !command.force {
         eprintln!("Already up-to-date.");
         return Ok(());
@@ -68,14 +69,14 @@ pub fn update(command: UpdateCommand) -> StrResult<()> {
 
     let binary_data = release.download_binary(needed_asset()?)?;
     let mut temp_exe = NamedTempFile::new()
-        .map_err(|err| eco_format!("failed to create temporary file: {err}"))?;
+        .map_err(|err| eco_format!("failed to create temporary file ({err})"))?;
     temp_exe
         .write_all(&binary_data)
-        .map_err(|err| eco_format!("failed to write binary data: {err}"))?;
+        .map_err(|err| eco_format!("failed to write binary data ({err})"))?;
 
     self_replace::self_replace(&temp_exe).map_err(|err| {
         fs::remove_file(&temp_exe).ok();
-        eco_format!("failed to self-replace running executable: {err}")
+        eco_format!("failed to self-replace running executable ({err})")
     })
 }
 
@@ -98,26 +99,24 @@ struct Release {
 impl Release {
     /// Download the target release, or latest if version is `None`, from the
     /// Typst repository.
-    pub fn from_tag(tag: Option<Version>) -> StrResult<Release> {
+    pub fn from_tag(tag: Option<&Version>) -> StrResult<Release> {
         let url = match tag {
             Some(tag) => format!(
-                "https://api.github.com/repos/{}/{}/releases/tags/v{}",
-                TYPST_GITHUB_ORG, TYPST_REPO, tag
+                "https://api.github.com/repos/{TYPST_GITHUB_ORG}/{TYPST_REPO}/releases/tags/v{tag}"
             ),
             None => format!(
-                "https://api.github.com/repos/{}/{}/releases/latest",
-                TYPST_GITHUB_ORG, TYPST_REPO
+                "https://api.github.com/repos/{TYPST_GITHUB_ORG}/{TYPST_REPO}/releases/latest",
             ),
         };
 
-        match ureq::get(&url).call() {
-            Ok(response) => response
-                .into_json()
-                .map_err(|err| eco_format!("unable to parse JSON response: {err}")),
+        match download(&url) {
+            Ok(response) => response.into_json().map_err(|err| {
+                eco_format!("failed to parse release information ({err})")
+            }),
             Err(ureq::Error::Status(404, _)) => {
                 bail!("release not found (searched at {url})")
             }
-            Err(_) => bail!("failed to download release (network failed)"),
+            Err(err) => bail!("failed to download release ({err})"),
         }
     }
 
@@ -132,19 +131,13 @@ impl Release {
             .ok_or("could not find release for your target platform")?;
 
         eprintln!("Downloading release ...");
-        let response = match ureq::get(&asset.browser_download_url).call() {
-            Ok(response) => response,
+        let data = match download_with_progress(&asset.browser_download_url) {
+            Ok(data) => data,
             Err(ureq::Error::Status(404, _)) => {
                 bail!("asset not found (searched for {})", asset.name);
             }
-            Err(_) => bail!("failed to load asset (network failed)"),
+            Err(err) => bail!("failed to download asset ({err})"),
         };
-
-        let mut data = Vec::new();
-        response
-            .into_reader()
-            .read_to_end(&mut data)
-            .map_err(|err| eco_format!("failed to read response buffer: {err}"))?;
 
         if asset_name.contains("windows") {
             extract_binary_from_zip(&data, asset_name)
@@ -157,15 +150,16 @@ impl Release {
 /// Extract the Typst binary from a ZIP archive.
 fn extract_binary_from_zip(data: &[u8], asset_name: &str) -> StrResult<Vec<u8>> {
     let mut archive = ZipArchive::new(Cursor::new(data))
-        .map_err(|err| eco_format!("failed to extract ZIP archive: {err}"))?;
+        .map_err(|err| eco_format!("failed to extract ZIP archive ({err})"))?;
 
-    let mut file = archive
-        .by_name(&format!("{asset_name}/typst.exe"))
-        .map_err(|_| "ZIP archive did not contain Typst binary")?;
+    let mut file =
+        archive.by_name(&format!("{asset_name}/typst.exe")).map_err(|err| {
+            eco_format!("failed to extract Typst binary from ZIP archive ({err})")
+        })?;
 
     let mut buffer = vec![];
     file.read_to_end(&mut buffer).map_err(|err| {
-        eco_format!("failed to read binary data from ZIP archive: {err}")
+        eco_format!("failed to read binary data from ZIP archive ({err})")
     })?;
 
     Ok(buffer)
@@ -177,14 +171,14 @@ fn extract_binary_from_tar_xz(data: &[u8]) -> StrResult<Vec<u8>> {
 
     let mut file = archive
         .entries()
-        .map_err(|err| eco_format!("failed to extract tar.xz archive: {err}"))?
+        .map_err(|err| eco_format!("failed to extract tar.xz archive ({err})"))?
         .filter_map(Result::ok)
         .find(|e| e.path().unwrap_or_default().ends_with("typst"))
         .ok_or("tar.xz archive did not contain Typst binary")?;
 
     let mut buffer = vec![];
     file.read_to_end(&mut buffer).map_err(|err| {
-        eco_format!("failed to read binary data from tar.xz archive: {err}")
+        eco_format!("failed to read binary data from tar.xz archive ({err})")
     })?;
 
     Ok(buffer)
@@ -230,7 +224,7 @@ fn update_needed(release: &Release) -> StrResult<bool> {
 fn backup_path() -> StrResult<PathBuf> {
     #[cfg(target_os = "linux")]
     let root_backup_dir = dirs::state_dir()
-        .or_else(|| dirs::data_dir())
+        .or_else(dirs::data_dir)
         .ok_or("unable to locate local data or state directory")?;
 
     #[cfg(not(target_os = "linux"))]
@@ -240,7 +234,7 @@ fn backup_path() -> StrResult<PathBuf> {
     let backup_dir = root_backup_dir.join("typst");
 
     fs::create_dir_all(&backup_dir)
-        .map_err(|err| eco_format!("failed to create backup directory: {err}"))?;
+        .map_err(|err| eco_format!("failed to create backup directory ({err})"))?;
 
     Ok(backup_dir.join("typst_backup.part"))
 }
