@@ -1,22 +1,29 @@
-use std::cell::{OnceCell, RefCell, RefMut};
 use std::collections::HashMap;
-use std::fs;
-use std::hash::Hash;
+use std::io::Read;
 use std::path::{Path, PathBuf};
+use std::sync::OnceLock;
+use std::{fmt, fs, io, mem};
 
-use chrono::{DateTime, Datelike, Local};
+use chrono::{DateTime, Datelike, FixedOffset, Local, Utc};
 use comemo::Prehashed;
-use same_file::Handle;
-use siphasher::sip128::{Hasher128, SipHasher13};
-use typst::diag::{FileError, FileResult, StrResult};
-use typst::eval::{eco_format, Bytes, Datetime, Library};
-use typst::font::{Font, FontBook};
+use ecow::{eco_format, EcoString};
+use once_cell::sync::Lazy;
+use parking_lot::Mutex;
+use typst::diag::{FileError, FileResult};
+use typst::foundations::{Bytes, Datetime, Dict, IntoValue};
 use typst::syntax::{FileId, Source, VirtualPath};
-use typst::World;
+use typst::text::{Font, FontBook};
+use typst::{Library, World};
+use typst_timing::{timed, TimingScope};
 
-use crate::args::SharedArgs;
+use crate::args::{Input, SharedArgs};
+use crate::compile::ExportCache;
 use crate::fonts::{FontSearcher, FontSlot};
-use crate::package::prepare_package;
+
+/// Static `FileId` allocated for stdin.
+/// This is to ensure that a file is read in the correct way.
+static STDIN_ID: Lazy<FileId> =
+    Lazy::new(|| FileId::new_fake(VirtualPath::new("<stdin>")));
 
 /// A world that provides access to the operating system.
 pub struct SystemWorld {
@@ -32,54 +39,87 @@ pub struct SystemWorld {
     book: Prehashed<FontBook>,
     /// Locations of and storage for lazily loaded fonts.
     fonts: Vec<FontSlot>,
-    /// Maps package-path combinations to canonical hashes. All package-path
-    /// combinations that point to the same file are mapped to the same hash. To
-    /// be used in conjunction with `paths`.
-    hashes: RefCell<HashMap<FileId, FileResult<PathHash>>>,
-    /// Maps canonical path hashes to source files and buffers.
-    paths: RefCell<HashMap<PathHash, PathSlot>>,
+    /// Maps file ids to source files and buffers.
+    slots: Mutex<HashMap<FileId, FileSlot>>,
     /// The current datetime if requested. This is stored here to ensure it is
-    /// always the same within one compilation. Reset between compilations.
-    now: OnceCell<DateTime<Local>>,
+    /// always the same within one compilation.
+    /// Reset between compilations if not [`Now::Fixed`].
+    now: Now,
+    /// The export cache, used for caching output files in `typst watch`
+    /// sessions.
+    export_cache: ExportCache,
 }
 
 impl SystemWorld {
     /// Create a new system world.
-    pub fn new(command: &SharedArgs) -> StrResult<Self> {
-        let mut searcher = FontSearcher::new();
-        searcher.search(&command.font_paths);
-
+    pub fn new(command: &SharedArgs) -> Result<Self, WorldCreationError> {
         // Resolve the system-global input path.
-        let input = command.input.canonicalize().map_err(|_| {
-            eco_format!("input file not found (searched at {})", command.input.display())
-        })?;
+        let input = match &command.input {
+            Input::Stdin => None,
+            Input::Path(path) => {
+                Some(path.canonicalize().map_err(|err| match err.kind() {
+                    io::ErrorKind::NotFound => {
+                        WorldCreationError::InputNotFound(path.clone())
+                    }
+                    _ => WorldCreationError::Io(err),
+                })?)
+            }
+        };
 
         // Resolve the system-global root directory.
         let root = {
             let path = command
                 .root
                 .as_deref()
-                .or_else(|| input.parent())
+                .or_else(|| input.as_deref().and_then(|i| i.parent()))
                 .unwrap_or(Path::new("."));
-            path.canonicalize().map_err(|_| {
-                eco_format!("root directory not found (searched at {})", path.display())
+            path.canonicalize().map_err(|err| match err.kind() {
+                io::ErrorKind::NotFound => {
+                    WorldCreationError::RootNotFound(path.to_path_buf())
+                }
+                _ => WorldCreationError::Io(err),
             })?
         };
 
-        // Resolve the virtual path of the main file within the project root.
-        let main_path = VirtualPath::within_root(&input, &root)
-            .ok_or("input file must be contained in project root")?;
+        let main = if let Some(path) = &input {
+            // Resolve the virtual path of the main file within the project root.
+            let main_path = VirtualPath::within_root(path, &root)
+                .ok_or(WorldCreationError::InputOutsideRoot)?;
+            FileId::new(None, main_path)
+        } else {
+            // Return the special id of STDIN otherwise
+            *STDIN_ID
+        };
+
+        let library = {
+            // Convert the input pairs to a dictionary.
+            let inputs: Dict = command
+                .inputs
+                .iter()
+                .map(|(k, v)| (k.as_str().into(), v.as_str().into_value()))
+                .collect();
+
+            Library::builder().with_inputs(inputs).build()
+        };
+
+        let mut searcher = FontSearcher::new();
+        searcher.search(&command.font_paths);
+
+        let now = match command.creation_timestamp {
+            Some(time) => Now::Fixed(time),
+            None => Now::System(OnceLock::new()),
+        };
 
         Ok(Self {
             workdir: std::env::current_dir().ok(),
             root,
-            main: FileId::new(None, main_path),
-            library: Prehashed::new(typst_library::build()),
+            main,
+            library: Prehashed::new(library),
             book: Prehashed::new(searcher.book),
             fonts: searcher.fonts,
-            hashes: RefCell::default(),
-            paths: RefCell::default(),
-            now: OnceCell::new(),
+            slots: Mutex::new(HashMap::new()),
+            now,
+            export_cache: ExportCache::new(),
         })
     }
 
@@ -99,21 +139,33 @@ impl SystemWorld {
     }
 
     /// Return all paths the last compilation depended on.
-    pub fn dependencies(&mut self) -> impl Iterator<Item = &Path> {
-        self.paths.get_mut().values().map(|slot| slot.system_path.as_path())
+    pub fn dependencies(&mut self) -> impl Iterator<Item = PathBuf> + '_ {
+        self.slots
+            .get_mut()
+            .values()
+            .filter(|slot| slot.accessed())
+            .filter_map(|slot| system_path(&self.root, slot.id).ok())
     }
 
     /// Reset the compilation state in preparation of a new compilation.
     pub fn reset(&mut self) {
-        self.hashes.borrow_mut().clear();
-        self.paths.borrow_mut().clear();
-        self.now.take();
+        for slot in self.slots.get_mut().values_mut() {
+            slot.reset();
+        }
+        if let Now::System(time_lock) = &mut self.now {
+            time_lock.take();
+        }
     }
 
     /// Lookup a source file by id.
     #[track_caller]
     pub fn lookup(&self, id: FileId) -> Source {
         self.source(id).expect("file id does not point to any source file")
+    }
+
+    /// Gets access to the export cache.
+    pub fn export_cache(&self) -> &ExportCache {
+        &self.export_cache
     }
 }
 
@@ -131,11 +183,11 @@ impl World for SystemWorld {
     }
 
     fn source(&self, id: FileId) -> FileResult<Source> {
-        self.slot(id)?.source()
+        self.slot(id, |slot| slot.source(&self.root))
     }
 
     fn file(&self, id: FileId) -> FileResult<Bytes> {
-        self.slot(id)?.file()
+        self.slot(id, |slot| slot.file(&self.root))
     }
 
     fn font(&self, index: usize) -> Option<Font> {
@@ -143,109 +195,184 @@ impl World for SystemWorld {
     }
 
     fn today(&self, offset: Option<i64>) -> Option<Datetime> {
-        let now = self.now.get_or_init(chrono::Local::now);
+        let now = match &self.now {
+            Now::Fixed(time) => time,
+            Now::System(time) => time.get_or_init(Utc::now),
+        };
 
-        let naive = match offset {
-            None => now.naive_local(),
-            Some(o) => now.naive_utc() + chrono::Duration::hours(o),
+        // The time with the specified UTC offset, or within the local time zone.
+        let with_offset = match offset {
+            None => now.with_timezone(&Local).fixed_offset(),
+            Some(hours) => {
+                let seconds = i32::try_from(hours).ok()?.checked_mul(3600)?;
+                now.with_timezone(&FixedOffset::east_opt(seconds)?)
+            }
         };
 
         Datetime::from_ymd(
-            naive.year(),
-            naive.month().try_into().ok()?,
-            naive.day().try_into().ok()?,
+            with_offset.year(),
+            with_offset.month().try_into().ok()?,
+            with_offset.day().try_into().ok()?,
         )
     }
 }
 
 impl SystemWorld {
     /// Access the canonical slot for the given file id.
-    #[tracing::instrument(skip_all)]
-    fn slot(&self, id: FileId) -> FileResult<RefMut<PathSlot>> {
-        let mut system_path = PathBuf::new();
-        let hash = self
-            .hashes
-            .borrow_mut()
-            .entry(id)
-            .or_insert_with(|| {
-                // Determine the root path relative to which the file path
-                // will be resolved.
-                let buf;
-                let mut root = &self.root;
-                if let Some(spec) = id.package() {
-                    buf = prepare_package(spec)?;
-                    root = &buf;
-                }
-
-                // Join the path to the root. If it tries to escape, deny
-                // access. Note: It can still escape via symlinks.
-                system_path = id.vpath().resolve(root).ok_or(FileError::AccessDenied)?;
-
-                PathHash::new(&system_path)
-            })
-            .clone()?;
-
-        Ok(RefMut::map(self.paths.borrow_mut(), |paths| {
-            paths.entry(hash).or_insert_with(|| PathSlot {
-                id,
-                // This will only trigger if the `or_insert_with` above also
-                // triggered.
-                system_path,
-                source: OnceCell::new(),
-                buffer: OnceCell::new(),
-            })
-        }))
+    fn slot<F, T>(&self, id: FileId, f: F) -> T
+    where
+        F: FnOnce(&mut FileSlot) -> T,
+    {
+        let mut map = self.slots.lock();
+        f(map.entry(id).or_insert_with(|| FileSlot::new(id)))
     }
 }
 
-/// Holds canonical data for all paths pointing to the same entity.
+/// Holds the processed data for a file ID.
 ///
 /// Both fields can be populated if the file is both imported and read().
-struct PathSlot {
-    /// The slot's canonical file id.
+struct FileSlot {
+    /// The slot's file id.
     id: FileId,
-    /// The slot's path on the system.
-    system_path: PathBuf,
-    /// The lazily loaded source file for a path hash.
-    source: OnceCell<FileResult<Source>>,
-    /// The lazily loaded buffer for a path hash.
-    buffer: OnceCell<FileResult<Bytes>>,
+    /// The lazily loaded and incrementally updated source file.
+    source: SlotCell<Source>,
+    /// The lazily loaded raw byte buffer.
+    file: SlotCell<Bytes>,
 }
 
-impl PathSlot {
-    fn source(&self) -> FileResult<Source> {
-        self.source
-            .get_or_init(|| {
-                let buf = read(&self.system_path)?;
-                let text = decode_utf8(buf)?;
-                Ok(Source::new(self.id, text))
-            })
-            .clone()
+impl FileSlot {
+    /// Create a new file slot.
+    fn new(id: FileId) -> Self {
+        Self { id, file: SlotCell::new(), source: SlotCell::new() }
     }
 
-    fn file(&self) -> FileResult<Bytes> {
-        self.buffer
-            .get_or_init(|| read(&self.system_path).map(Bytes::from))
-            .clone()
+    /// Whether the file was accessed in the ongoing compilation.
+    fn accessed(&self) -> bool {
+        self.source.accessed() || self.file.accessed()
+    }
+
+    /// Marks the file as not yet accessed in preparation of the next
+    /// compilation.
+    fn reset(&mut self) {
+        self.source.reset();
+        self.file.reset();
+    }
+
+    /// Retrieve the source for this file.
+    fn source(&mut self, project_root: &Path) -> FileResult<Source> {
+        self.source.get_or_init(
+            || read(self.id, project_root),
+            |data, prev| {
+                let name = if prev.is_some() { "reparsing file" } else { "parsing file" };
+                let _scope = TimingScope::new(name, None);
+                let text = decode_utf8(&data)?;
+                if let Some(mut prev) = prev {
+                    prev.replace(text);
+                    Ok(prev)
+                } else {
+                    Ok(Source::new(self.id, text.into()))
+                }
+            },
+        )
+    }
+
+    /// Retrieve the file's bytes.
+    fn file(&mut self, project_root: &Path) -> FileResult<Bytes> {
+        self.file
+            .get_or_init(|| read(self.id, project_root), |data, _| Ok(data.into()))
     }
 }
 
-/// A hash that is the same for all paths pointing to the same entity.
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-struct PathHash(u128);
+/// Lazily processes data for a file.
+struct SlotCell<T> {
+    /// The processed data.
+    data: Option<FileResult<T>>,
+    /// A hash of the raw file contents / access error.
+    fingerprint: u128,
+    /// Whether the slot has been accessed in the current compilation.
+    accessed: bool,
+}
 
-impl PathHash {
-    fn new(path: &Path) -> FileResult<Self> {
-        let f = |e| FileError::from_io(e, path);
-        let handle = Handle::from_path(path).map_err(f)?;
-        let mut state = SipHasher13::new();
-        handle.hash(&mut state);
-        Ok(Self(state.finish128().as_u128()))
+impl<T: Clone> SlotCell<T> {
+    /// Creates a new, empty cell.
+    fn new() -> Self {
+        Self { data: None, fingerprint: 0, accessed: false }
+    }
+
+    /// Whether the cell was accessed in the ongoing compilation.
+    fn accessed(&self) -> bool {
+        self.accessed
+    }
+
+    /// Marks the cell as not yet accessed in preparation of the next
+    /// compilation.
+    fn reset(&mut self) {
+        self.accessed = false;
+    }
+
+    /// Gets the contents of the cell or initialize them.
+    fn get_or_init(
+        &mut self,
+        load: impl FnOnce() -> FileResult<Vec<u8>>,
+        f: impl FnOnce(Vec<u8>, Option<T>) -> FileResult<T>,
+    ) -> FileResult<T> {
+        // If we accessed the file already in this compilation, retrieve it.
+        if mem::replace(&mut self.accessed, true) {
+            if let Some(data) = &self.data {
+                return data.clone();
+            }
+        }
+
+        // Read and hash the file.
+        let result = timed!("loading file", load());
+        let fingerprint = timed!("hashing file", typst::util::hash128(&result));
+
+        // If the file contents didn't change, yield the old processed data.
+        if mem::replace(&mut self.fingerprint, fingerprint) == fingerprint {
+            if let Some(data) = &self.data {
+                return data.clone();
+            }
+        }
+
+        let prev = self.data.take().and_then(Result::ok);
+        let value = result.and_then(|data| f(data, prev));
+        self.data = Some(value.clone());
+
+        value
     }
 }
 
-/// Read a file.
-fn read(path: &Path) -> FileResult<Vec<u8>> {
+/// Resolves the path of a file id on the system, downloading a package if
+/// necessary.
+fn system_path(project_root: &Path, id: FileId) -> FileResult<PathBuf> {
+    // Determine the root path relative to which the file path
+    // will be resolved.
+    let buf;
+    let mut root = project_root;
+    if let Some(spec) = id.package() {
+        buf = crate::package::prepare_package(spec)?;
+        root = &buf;
+    }
+
+    // Join the path to the root. If it tries to escape, deny
+    // access. Note: It can still escape via symlinks.
+    id.vpath().resolve(root).ok_or(FileError::AccessDenied)
+}
+
+/// Reads a file from a `FileId`.
+///
+/// If the ID represents stdin it will read from standard input,
+/// otherwise it gets the file path of the ID and reads the file from disk.
+fn read(id: FileId, project_root: &Path) -> FileResult<Vec<u8>> {
+    if id == *STDIN_ID {
+        read_from_stdin()
+    } else {
+        read_from_disk(&system_path(project_root, id)?)
+    }
+}
+
+/// Read a file from disk.
+fn read_from_disk(path: &Path) -> FileResult<Vec<u8>> {
     let f = |e| FileError::from_io(e, path);
     if fs::metadata(path).map_err(f)?.is_dir() {
         Err(FileError::IsDirectory)
@@ -254,13 +381,65 @@ fn read(path: &Path) -> FileResult<Vec<u8>> {
     }
 }
 
+/// Read from stdin.
+fn read_from_stdin() -> FileResult<Vec<u8>> {
+    let mut buf = Vec::new();
+    let result = io::stdin().read_to_end(&mut buf);
+    match result {
+        Ok(_) => (),
+        Err(err) if err.kind() == io::ErrorKind::BrokenPipe => (),
+        Err(err) => return Err(FileError::from_io(err, Path::new("<stdin>"))),
+    }
+    Ok(buf)
+}
+
 /// Decode UTF-8 with an optional BOM.
-fn decode_utf8(buf: Vec<u8>) -> FileResult<String> {
-    Ok(if buf.starts_with(b"\xef\xbb\xbf") {
-        // Remove UTF-8 BOM.
-        std::str::from_utf8(&buf[3..])?.into()
-    } else {
-        // Assume UTF-8.
-        String::from_utf8(buf)?
-    })
+fn decode_utf8(buf: &[u8]) -> FileResult<&str> {
+    // Remove UTF-8 BOM.
+    Ok(std::str::from_utf8(buf.strip_prefix(b"\xef\xbb\xbf").unwrap_or(buf))?)
+}
+
+/// The current date and time.
+enum Now {
+    /// The date and time if the environment `SOURCE_DATE_EPOCH` is set.
+    /// Used for reproducible builds.
+    Fixed(DateTime<Utc>),
+    /// The current date and time if the time is not externally fixed.
+    System(OnceLock<DateTime<Utc>>),
+}
+
+/// An error that occurs during world construction.
+#[derive(Debug)]
+pub enum WorldCreationError {
+    /// The input file does not appear to exist.
+    InputNotFound(PathBuf),
+    /// The input file is not contained within the root folder.
+    InputOutsideRoot,
+    /// The root directory does not appear to exist.
+    RootNotFound(PathBuf),
+    /// Another type of I/O error.
+    Io(io::Error),
+}
+
+impl fmt::Display for WorldCreationError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            WorldCreationError::InputNotFound(path) => {
+                write!(f, "input file not found (searched at {})", path.display())
+            }
+            WorldCreationError::InputOutsideRoot => {
+                write!(f, "source file must be contained in project root")
+            }
+            WorldCreationError::RootNotFound(path) => {
+                write!(f, "root directory not found (searched at {})", path.display())
+            }
+            WorldCreationError::Io(err) => write!(f, "{err}"),
+        }
+    }
+}
+
+impl From<WorldCreationError> for EcoString {
+    fn from(err: WorldCreationError) -> Self {
+        eco_format!("{err}")
+    }
 }
