@@ -1,23 +1,17 @@
 //! File and package management.
 
 use std::collections::HashMap;
-use std::fmt::{self, Debug, Display, Formatter};
-use std::path::{Component, Path, PathBuf};
-use std::str::FromStr;
+use std::fmt::{self, Debug, Formatter};
 use std::sync::RwLock;
 
-use ecow::{eco_format, EcoString};
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
-use super::is_ident;
+use crate::package::PackageSpec;
+use crate::VirtualPath;
 
 /// The global package-path interner.
 static INTERNER: Lazy<RwLock<Interner>> =
     Lazy::new(|| RwLock::new(Interner { to_id: HashMap::new(), from_id: Vec::new() }));
-
-/// The path that we use for detached file ids.
-static DETACHED_PATH: Lazy<VirtualPath> = Lazy::new(|| VirtualPath::new("/unknown"));
 
 /// A package-path interner.
 struct Interner {
@@ -48,66 +42,59 @@ impl FileId {
         }
 
         let mut interner = INTERNER.write().unwrap();
-        let len = interner.from_id.len();
-        if len >= usize::from(u16::MAX) {
-            panic!("too many file specifications");
-        }
+        let num = interner.from_id.len().try_into().expect("out of file ids");
 
         // Create a new entry forever by leaking the pair. We can't leak more
         // than 2^16 pair (and typically will leak a lot less), so its not a
         // big deal.
-        let id = FileId(len as u16);
+        let id = FileId(num);
         let leaked = Box::leak(Box::new(pair));
         interner.to_id.insert(leaked, id);
         interner.from_id.push(leaked);
         id
     }
 
-    /// Get an id that does not identify any real file.
-    pub const fn detached() -> Self {
-        Self(u16::MAX)
-    }
+    /// Create a new unique ("fake") file specification, which is not
+    /// accessible by path.
+    ///
+    /// Caution: the ID returned by this method is the *only* identifier of the
+    /// file, constructing a file ID with a path will *not* reuse the ID even
+    /// if the path is the same. This method should only be used for generating
+    /// "virtual" file ids such as content read from stdin.
+    #[track_caller]
+    pub fn new_fake(path: VirtualPath) -> Self {
+        let mut interner = INTERNER.write().unwrap();
+        let num = interner.from_id.len().try_into().expect("out of file ids");
 
-    /// Whether the id is the detached.
-    pub const fn is_detached(self) -> bool {
-        self.0 == Self::detached().0
+        let id = FileId(num);
+        let leaked = Box::leak(Box::new((None, path)));
+        interner.from_id.push(leaked);
+        id
     }
 
     /// The package the file resides in, if any.
     pub fn package(&self) -> Option<&'static PackageSpec> {
-        if self.is_detached() {
-            None
-        } else {
-            self.pair().0.as_ref()
-        }
+        self.pair().0.as_ref()
     }
 
     /// The absolute and normalized path to the file _within_ the project or
     /// package.
     pub fn vpath(&self) -> &'static VirtualPath {
-        if self.is_detached() {
-            &DETACHED_PATH
-        } else {
-            &self.pair().1
-        }
+        &self.pair().1
     }
 
     /// Resolve a file location relative to this file.
-    pub fn join(self, path: &str) -> Result<Self, EcoString> {
-        if self.is_detached() {
-            Err("cannot access file system from here")?;
-        }
-
-        Ok(Self::new(self.package().cloned(), self.vpath().join(path)))
+    pub fn join(self, path: &str) -> Self {
+        Self::new(self.package().cloned(), self.vpath().join(path))
     }
 
     /// Construct from a raw number.
-    pub(crate) const fn from_u16(v: u16) -> Self {
+    pub(crate) const fn from_raw(v: u16) -> Self {
         Self(v)
     }
 
     /// Extract the raw underlying number.
-    pub(crate) const fn as_u16(self) -> u16 {
+    pub(crate) const fn into_raw(self) -> u16 {
         self.0
     }
 
@@ -124,221 +111,5 @@ impl Debug for FileId {
             Some(package) => write!(f, "{package:?}{vpath:?}"),
             None => write!(f, "{vpath:?}"),
         }
-    }
-}
-
-/// An absolute path in the virtual file system of a project or package.
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct VirtualPath(PathBuf);
-
-impl VirtualPath {
-    /// Create a new virtual path.
-    ///
-    /// Even if it doesn't start with `/` or `\`, it is still interpreted as
-    /// starting from the root.
-    pub fn new(path: impl AsRef<Path>) -> Self {
-        Self::new_impl(path.as_ref())
-    }
-
-    /// Non generic new implementation.
-    fn new_impl(path: &Path) -> Self {
-        let mut out = Path::new(&Component::RootDir).to_path_buf();
-        for component in path.components() {
-            match component {
-                Component::Prefix(_) | Component::RootDir => {}
-                Component::CurDir => {}
-                Component::ParentDir => match out.components().next_back() {
-                    Some(Component::Normal(_)) => {
-                        out.pop();
-                    }
-                    _ => out.push(component),
-                },
-                Component::Normal(_) => out.push(component),
-            }
-        }
-        Self(out)
-    }
-
-    /// Create a virtual path from a real path and a real root.
-    ///
-    /// Returns `None` if the file path is not contained in the root (i.e. if
-    /// `root` is not a lexical prefix of `path`). No file system operations are
-    /// performed.
-    pub fn within_root(path: &Path, root: &Path) -> Option<Self> {
-        path.strip_prefix(root).ok().map(Self::new)
-    }
-
-    /// Get the underlying path with a leading `/` or `\`.
-    pub fn as_rooted_path(&self) -> &Path {
-        &self.0
-    }
-
-    /// Get the underlying path without a leading `/` or `\`.
-    pub fn as_rootless_path(&self) -> &Path {
-        self.0.strip_prefix(Component::RootDir).unwrap_or(&self.0)
-    }
-
-    /// Resolve the virtual path relative to an actual file system root
-    /// (where the project or package resides).
-    ///
-    /// Returns `None` if the path lexically escapes the root. The path might
-    /// still escape through symlinks.
-    pub fn resolve(&self, root: &Path) -> Option<PathBuf> {
-        let root_len = root.as_os_str().len();
-        let mut out = root.to_path_buf();
-        for component in self.0.components() {
-            match component {
-                Component::Prefix(_) => {}
-                Component::RootDir => {}
-                Component::CurDir => {}
-                Component::ParentDir => {
-                    out.pop();
-                    if out.as_os_str().len() < root_len {
-                        return None;
-                    }
-                }
-                Component::Normal(_) => out.push(component),
-            }
-        }
-        Some(out)
-    }
-
-    /// Resolve a path relative to this virtual path.
-    pub fn join(&self, path: impl AsRef<Path>) -> Self {
-        if let Some(parent) = self.0.parent() {
-            Self::new(parent.join(path))
-        } else {
-            Self::new(path)
-        }
-    }
-}
-
-impl Debug for VirtualPath {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        Display::fmt(&self.0.display(), f)
-    }
-}
-
-/// Identifies a package.
-#[derive(Clone, Eq, PartialEq, Hash)]
-pub struct PackageSpec {
-    /// The namespace the package lives in.
-    pub namespace: EcoString,
-    /// The name of the package within its namespace.
-    pub name: EcoString,
-    /// The package's version.
-    pub version: PackageVersion,
-}
-
-impl FromStr for PackageSpec {
-    type Err = EcoString;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut s = unscanny::Scanner::new(s);
-        if !s.eat_if('@') {
-            Err("package specification must start with '@'")?;
-        }
-
-        let namespace = s.eat_until('/');
-        if namespace.is_empty() {
-            Err("package specification is missing namespace")?;
-        } else if !is_ident(namespace) {
-            Err(eco_format!("`{namespace}` is not a valid package namespace"))?;
-        }
-
-        s.eat_if('/');
-
-        let name = s.eat_until(':');
-        if name.is_empty() {
-            Err("package specification is missing name")?;
-        } else if !is_ident(name) {
-            Err(eco_format!("`{name}` is not a valid package name"))?;
-        }
-
-        s.eat_if(':');
-
-        let version = s.after();
-        if version.is_empty() {
-            Err("package specification is missing version")?;
-        }
-
-        Ok(Self {
-            namespace: namespace.into(),
-            name: name.into(),
-            version: version.parse()?,
-        })
-    }
-}
-
-impl Debug for PackageSpec {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        Display::fmt(self, f)
-    }
-}
-
-impl Display for PackageSpec {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "@{}/{}:{}", self.namespace, self.name, self.version)
-    }
-}
-
-/// A package's version.
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct PackageVersion {
-    /// The package's major version.
-    pub major: u32,
-    /// The package's minor version.
-    pub minor: u32,
-    /// The package's patch version.
-    pub patch: u32,
-}
-
-impl FromStr for PackageVersion {
-    type Err = EcoString;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut parts = s.split('.');
-        let mut next = |kind| {
-            let part = parts
-                .next()
-                .filter(|s| !s.is_empty())
-                .ok_or_else(|| eco_format!("version number is missing {kind} version"))?;
-            part.parse::<u32>()
-                .map_err(|_| eco_format!("`{part}` is not a valid {kind} version"))
-        };
-
-        let major = next("major")?;
-        let minor = next("minor")?;
-        let patch = next("patch")?;
-        if let Some(rest) = parts.next() {
-            Err(eco_format!("version number has unexpected fourth component: `{rest}`"))?;
-        }
-
-        Ok(Self { major, minor, patch })
-    }
-}
-
-impl Debug for PackageVersion {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        Display::fmt(self, f)
-    }
-}
-
-impl Display for PackageVersion {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "{}.{}.{}", self.major, self.minor, self.patch)
-    }
-}
-
-impl Serialize for PackageVersion {
-    fn serialize<S: Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
-        s.collect_str(self)
-    }
-}
-
-impl<'de> Deserialize<'de> for PackageVersion {
-    fn deserialize<D: Deserializer<'de>>(d: D) -> Result<Self, D::Error> {
-        let string = EcoString::deserialize(d)?;
-        string.parse().map_err(serde::de::Error::custom)
     }
 }
