@@ -4,7 +4,7 @@
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 #[repr(u8)]
 pub enum SyntaxKind {
-    /// Markup.
+    /// The contents of a file or content block.
     Markup,
     /// Plain text without markup.
     Text,
@@ -28,6 +28,12 @@ pub enum SyntaxKind {
     Emph,
     /// Raw text with optional syntax highlighting: `` `...` ``.
     Raw,
+    /// A language tag at the start of raw text: ``typ ``.
+    RawLang,
+    /// A raw delimiter consisting of 1 or 3+ backticks: `` ` ``.
+    RawDelim,
+    /// A sequence of whitespace to ignore in a raw text: `    `.
+    RawTrimmed,
     /// A hyperlink: `https://typst.org`.
     Link,
     /// A label: `<intro>`.
@@ -72,8 +78,8 @@ pub enum SyntaxKind {
     /// A root in math: `√x`, `∛x` or `∜x`.
     MathRoot,
 
-    /// A hashtag that switches into code mode: `#`.
-    Hashtag,
+    /// A hash that switches into code mode: `#`.
+    Hash,
     /// A left curly brace, starting a code block: `{`.
     LeftBrace,
     /// A right curly brace, terminating a code block: `}`.
@@ -136,7 +142,7 @@ pub enum SyntaxKind {
     StarEq,
     /// The divide-assign operator: `/=`.
     SlashEq,
-    /// The spread operator: `..`.
+    /// Indicates a spread or sink: `..`.
     Dots,
     /// An arrow between a closure's parameters and body: `=>`.
     Arrow,
@@ -159,6 +165,8 @@ pub enum SyntaxKind {
     Set,
     /// The `show` keyword.
     Show,
+    /// The `context` keyword.
+    Context,
     /// The `if` keyword.
     If,
     /// The `else` keyword.
@@ -182,7 +190,7 @@ pub enum SyntaxKind {
     /// The `as` keyword.
     As,
 
-    /// Code.
+    /// The contents of a code block.
     Code,
     /// An identifier: `it`.
     Ident,
@@ -232,6 +240,8 @@ pub enum SyntaxKind {
     SetRule,
     /// A show rule: `show heading: it => emph(it.body)`.
     ShowRule,
+    /// A contextual expression: `context text.lang`.
+    Contextual,
     /// An if-else conditional: `if x { y } else { z }`.
     Conditional,
     /// A while loop: `while x { y }`.
@@ -242,6 +252,8 @@ pub enum SyntaxKind {
     ModuleImport,
     /// Items to import from a module: `a, b, c`.
     ImportItems,
+    /// A renamed import item: `a as d`.
+    RenamedImportItem,
     /// A module include: `include "chapter1.typ"`.
     ModuleInclude,
     /// A break from a loop: `break`.
@@ -261,8 +273,8 @@ pub enum SyntaxKind {
     BlockComment,
     /// An invalid sequence of characters.
     Error,
-    /// The end of the file.
-    Eof,
+    /// The end of token stream.
+    End,
 }
 
 impl SyntaxKind {
@@ -283,7 +295,7 @@ impl SyntaxKind {
     pub fn is_terminator(self) -> bool {
         matches!(
             self,
-            Self::Eof
+            Self::End
                 | Self::Semicolon
                 | Self::RightBrace
                 | Self::RightParen
@@ -320,6 +332,7 @@ impl SyntaxKind {
                 | Self::Let
                 | Self::Set
                 | Self::Show
+                | Self::Context
                 | Self::If
                 | Self::Else
                 | Self::For
@@ -362,6 +375,9 @@ impl SyntaxKind {
             Self::Strong => "strong content",
             Self::Emph => "emphasized content",
             Self::Raw => "raw block",
+            Self::RawLang => "raw language tag",
+            Self::RawTrimmed => "raw trimmed",
+            Self::RawDelim => "raw delimiter",
             Self::Link => "link",
             Self::Label => "label",
             Self::Ref => "reference",
@@ -383,7 +399,7 @@ impl SyntaxKind {
             Self::MathFrac => "math fraction",
             Self::MathRoot => "math root",
             Self::MathPrimes => "math primes",
-            Self::Hashtag => "hashtag",
+            Self::Hash => "hash",
             Self::LeftBrace => "opening brace",
             Self::RightBrace => "closing brace",
             Self::LeftBracket => "opening bracket",
@@ -424,6 +440,7 @@ impl SyntaxKind {
             Self::Let => "keyword `let`",
             Self::Set => "keyword `set`",
             Self::Show => "keyword `show`",
+            Self::Context => "keyword `context`",
             Self::If => "keyword `if`",
             Self::Else => "keyword `else`",
             Self::For => "keyword `for`",
@@ -460,11 +477,13 @@ impl SyntaxKind {
             Self::LetBinding => "`let` expression",
             Self::SetRule => "`set` expression",
             Self::ShowRule => "`show` expression",
+            Self::Contextual => "`context` expression",
             Self::Conditional => "`if` expression",
             Self::WhileLoop => "while-loop expression",
             Self::ForLoop => "for-loop expression",
             Self::ModuleImport => "`import` expression",
             Self::ImportItems => "import items",
+            Self::RenamedImportItem => "renamed import item",
             Self::ModuleInclude => "`include` expression",
             Self::LoopBreak => "`break` expression",
             Self::LoopContinue => "`continue` expression",
@@ -474,7 +493,7 @@ impl SyntaxKind {
             Self::LineComment => "line comment",
             Self::BlockComment => "block comment",
             Self::Error => "syntax error",
-            Self::Eof => "end of file",
+            Self::End => "end of tokens",
         }
     }
 }
