@@ -1,6 +1,6 @@
 use std::ops::Range;
 
-use super::{
+use crate::{
     is_newline, parse, reparse_block, reparse_markup, Span, SyntaxKind, SyntaxNode,
 };
 
@@ -11,7 +11,7 @@ use super::{
 /// ultimately reparsed.
 ///
 /// The high-level API for this function is
-/// [`Source::edit`](super::Source::edit).
+/// [`Source::edit`](crate::Source::edit).
 pub fn reparse(
     root: &mut SyntaxNode,
     text: &str,
@@ -21,7 +21,9 @@ pub fn reparse(
     try_reparse(text, replaced, replacement_len, None, root, 0).unwrap_or_else(|| {
         let id = root.span().id();
         *root = parse(text);
-        root.numberize(id, Span::FULL).unwrap();
+        if let Some(id) = id {
+            root.numberize(id, Span::FULL).unwrap();
+        }
         0..text.len()
     })
 }
@@ -123,8 +125,8 @@ fn try_reparse(
             end += 1;
         }
 
-        // Also take hashtag.
-        if start > 0 && children[start - 1].kind() == SyntaxKind::Hashtag {
+        // Also take hash.
+        if start > 0 && children[start - 1].kind() == SyntaxKind::Hash {
             start -= 1;
         }
 
@@ -158,7 +160,7 @@ fn try_reparse(
         // Stop parsing early if this kind is encountered.
         let stop_kind = match parent_kind {
             Some(_) => SyntaxKind::RightBracket,
-            None => SyntaxKind::Eof,
+            None => SyntaxKind::End,
         };
 
         // Reparse!
@@ -244,7 +246,7 @@ fn next_nesting(node: &SyntaxNode, nesting: &mut usize) {
 mod tests {
     use std::ops::Range;
 
-    use super::super::{parse, Source, Span};
+    use crate::{parse, Source, Span};
 
     #[track_caller]
     fn test(prev: &str, range: Range<usize>, with: &str, incremental: bool) {
@@ -299,6 +301,8 @@ mod tests {
         test("a\n#let \nb", 7..7, "i", true);
         test(r"#{{let x = z}; a = 1} b", 7..7, "//", false);
         test(r#"a ```typst hello```"#, 16..17, "", false);
+        test("a{b}c", 1..1, "#", false);
+        test("a#{b}c", 1..2, "", false);
     }
 
     #[test]
