@@ -7,9 +7,11 @@ use std::str::Utf8Error;
 use std::string::FromUtf8Error;
 
 use comemo::Tracked;
+use ecow::{eco_vec, EcoVec};
 
-use crate::syntax::{PackageSpec, Span, Spanned, SyntaxError};
-use crate::World;
+use crate::syntax::package::PackageSpec;
+use crate::syntax::{Span, Spanned, SyntaxError};
+use crate::{World, WorldExt};
 
 /// Early-return with a [`StrResult`] or [`SourceResult`].
 ///
@@ -17,69 +19,107 @@ use crate::World;
 /// `StrResult`. If called with a span, a string and format args, returns
 /// a `SourceResult`.
 ///
-/// ```
+/// You can also emit hints with the `; hint: "..."` syntax.
+///
+/// ```ignore
 /// bail!("bailing with a {}", "string result");
 /// bail!(span, "bailing with a {}", "source result");
+/// bail!(
+///     span, "bailing with a {}", "source result";
+///     hint: "hint 1"
+/// );
+/// bail!(
+///     span, "bailing with a {}", "source result";
+///     hint: "hint 1";
+///     hint: "hint 2";
+/// );
 /// ```
 #[macro_export]
 #[doc(hidden)]
 macro_rules! __bail {
+    // For bail!("just a {}", "string")
     ($fmt:literal $(, $arg:expr)* $(,)?) => {
-        return Err($crate::diag::eco_format!($fmt, $($arg),*))
+        return Err($crate::diag::error!(
+            $fmt, $($arg),*
+        ))
     };
 
+    // For bail!(error!(..))
     ($error:expr) => {
-        return Err(Box::new(vec![$error]))
+        return Err(::ecow::eco_vec![$error])
     };
 
-    ($span:expr, $fmt:literal $(, $arg:expr)* $(,)?) => {
-        return Err(Box::new(vec![$crate::diag::SourceDiagnostic::error(
-            $span,
-            $crate::diag::eco_format!($fmt, $($arg),*),
-        )]))
+    // For bail(span, ...)
+    ($($tts:tt)*) => {
+        return Err(::ecow::eco_vec![$crate::diag::error!($($tts)*)])
     };
 }
-
-#[doc(inline)]
-pub use crate::__bail as bail;
 
 /// Construct an [`EcoString`] or [`SourceDiagnostic`] with severity `Error`.
 #[macro_export]
 #[doc(hidden)]
 macro_rules! __error {
+    // For bail!("just a {}", "string").
     ($fmt:literal $(, $arg:expr)* $(,)?) => {
         $crate::diag::eco_format!($fmt, $($arg),*)
     };
 
-    ($span:expr, $fmt:literal $(, $arg:expr)* $(,)?) => {
+    // For bail!(span, ...)
+    (
+        $span:expr, $fmt:literal $(, $arg:expr)*
+        $(; hint: $hint:literal $(, $hint_arg:expr)*)*
+        $(,)?
+    ) => {
         $crate::diag::SourceDiagnostic::error(
             $span,
             $crate::diag::eco_format!($fmt, $($arg),*),
-        )
+        )  $(.with_hint($crate::diag::eco_format!($hint, $($hint_arg),*)))*
     };
 }
 
 /// Construct a [`SourceDiagnostic`] with severity `Warning`.
+///
+/// You can also emit hints with the `; hint: "..."` syntax.
+///
+/// ```ignore
+/// warning!(span, "warning with a {}", "source result");
+/// warning!(
+///     span, "warning with a {}", "source result";
+///     hint: "hint 1"
+/// );
+/// warning!(
+///     span, "warning with a {}", "source result";
+///     hint: "hint 1";
+///     hint: "hint 2";
+/// );
+/// ```
 #[macro_export]
 #[doc(hidden)]
 macro_rules! __warning {
-    ($span:expr, $fmt:literal $(, $arg:expr)* $(,)?) => {
+    (
+        $span:expr,
+        $fmt:literal $(, $arg:expr)*
+        $(; hint: $hint:literal $(, $hint_arg:expr)*)*
+        $(,)?
+    ) => {
         $crate::diag::SourceDiagnostic::warning(
             $span,
             $crate::diag::eco_format!($fmt, $($arg),*),
-        )
+        ) $(.with_hint($crate::diag::eco_format!($hint, $($hint_arg),*)))*
     };
 }
 
+#[rustfmt::skip]
 #[doc(inline)]
-pub use crate::__error as error;
-#[doc(inline)]
-pub use crate::__warning as warning;
-#[doc(hidden)]
-pub use ecow::{eco_format, EcoString};
+pub use {
+    crate::__bail as bail,
+    crate::__error as error,
+    crate::__warning as warning,
+    ecow::{eco_format, EcoString},
+};
 
 /// A result that can carry multiple source errors.
-pub type SourceResult<T> = Result<T, Box<Vec<SourceDiagnostic>>>;
+pub type SourceResult<T> = Result<T, EcoVec<SourceDiagnostic>>;
 
 /// An error or warning in a source file.
 ///
@@ -94,10 +134,10 @@ pub struct SourceDiagnostic {
     /// A diagnostic message describing the problem.
     pub message: EcoString,
     /// The trace of function calls leading to the problem.
-    pub trace: Vec<Spanned<Tracepoint>>,
-    /// Additonal hints to the user, indicating how this problem could be avoided
+    pub trace: EcoVec<Spanned<Tracepoint>>,
+    /// Additional hints to the user, indicating how this problem could be avoided
     /// or worked around.
-    pub hints: Vec<EcoString>,
+    pub hints: EcoVec<EcoString>,
 }
 
 /// The severity of a [`SourceDiagnostic`].
@@ -115,9 +155,9 @@ impl SourceDiagnostic {
         Self {
             severity: Severity::Error,
             span,
-            trace: vec![],
+            trace: eco_vec![],
             message: message.into(),
-            hints: vec![],
+            hints: eco_vec![],
         }
     }
 
@@ -126,9 +166,9 @@ impl SourceDiagnostic {
         Self {
             severity: Severity::Warning,
             span,
-            trace: vec![],
+            trace: eco_vec![],
             message: message.into(),
-            hints: vec![],
+            hints: eco_vec![],
         }
     }
 
@@ -156,7 +196,7 @@ impl From<SyntaxError> for SourceDiagnostic {
             severity: Severity::Error,
             span: error.span,
             message: error.message,
-            trace: vec![],
+            trace: eco_vec![],
             hints: error.hints,
         }
     }
@@ -177,7 +217,7 @@ impl Display for Tracepoint {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         match self {
             Tracepoint::Call(Some(name)) => {
-                write!(f, "error occurred in this call of function `{}`", name)
+                write!(f, "error occurred in this call of function `{name}`")
             }
             Tracepoint::Call(None) => {
                 write!(f, "error occurred in this function call")
@@ -206,16 +246,12 @@ impl<T> Trace<T> for SourceResult<T> {
         F: Fn() -> Tracepoint,
     {
         self.map_err(|mut errors| {
-            if span.is_detached() {
-                return errors;
-            }
-
-            let trace_range = world.range(span);
-            for error in errors.iter_mut().filter(|e| !e.span.is_detached()) {
+            let Some(trace_range) = world.range(span) else { return errors };
+            for error in errors.make_mut().iter_mut() {
                 // Skip traces that surround the error.
-                if error.span.id() == span.id() {
-                    let error_range = world.range(error.span);
-                    if trace_range.start <= error_range.start
+                if let Some(error_range) = world.range(error.span) {
+                    if error.span.id() == span.id()
+                        && trace_range.start <= error_range.start
                         && trace_range.end >= error_range.end
                     {
                         continue;
@@ -232,7 +268,8 @@ impl<T> Trace<T> for SourceResult<T> {
 /// A result type with a string error message.
 pub type StrResult<T> = Result<T, EcoString>;
 
-/// Convert a [`StrResult`] to a [`SourceResult`] by adding span information.
+/// Convert a [`StrResult`] or [`HintedStrResult`] to a [`SourceResult`] by
+/// adding span information.
 pub trait At<T> {
     /// Add the span information.
     fn at(self, span: Span) -> SourceResult<T>;
@@ -250,7 +287,7 @@ where
                 diagnostic
                     .hint("you can adjust the project root with the --root argument");
             }
-            Box::new(vec![diagnostic])
+            eco_vec![diagnostic]
         })
     }
 }
@@ -263,17 +300,24 @@ pub type HintedStrResult<T> = Result<T, HintedString>;
 pub struct HintedString {
     /// A diagnostic message describing the problem.
     pub message: EcoString,
-    /// Additonal hints to the user, indicating how this error could be avoided
+    /// Additional hints to the user, indicating how this error could be avoided
     /// or worked around.
     pub hints: Vec<EcoString>,
+}
+
+impl<S> From<S> for HintedString
+where
+    S: Into<EcoString>,
+{
+    fn from(value: S) -> Self {
+        Self { message: value.into(), hints: vec![] }
+    }
 }
 
 impl<T> At<T> for Result<T, HintedString> {
     fn at(self, span: Span) -> SourceResult<T> {
         self.map_err(|diags| {
-            Box::new(vec![
-                SourceDiagnostic::error(span, diags.message).with_hints(diags.hints)
-            ])
+            eco_vec![SourceDiagnostic::error(span, diags.message).with_hints(diags.hints)]
         })
     }
 }
@@ -324,21 +368,23 @@ pub enum FileError {
     /// The package the file is part of could not be loaded.
     Package(PackageError),
     /// Another error.
-    Other,
+    ///
+    /// The optional string can give more details, if available.
+    Other(Option<EcoString>),
 }
 
 impl FileError {
     /// Create a file error from an I/O error.
-    pub fn from_io(error: io::Error, path: &Path) -> Self {
-        match error.kind() {
+    pub fn from_io(err: io::Error, path: &Path) -> Self {
+        match err.kind() {
             io::ErrorKind::NotFound => Self::NotFound(path.into()),
             io::ErrorKind::PermissionDenied => Self::AccessDenied,
             io::ErrorKind::InvalidData
-                if error.to_string().contains("stream did not contain valid UTF-8") =>
+                if err.to_string().contains("stream did not contain valid UTF-8") =>
             {
                 Self::InvalidUtf8
             }
-            _ => Self::Other,
+            _ => Self::Other(Some(eco_format!("{err}"))),
         }
     }
 }
@@ -356,7 +402,8 @@ impl Display for FileError {
             Self::NotSource => f.pad("not a typst source file"),
             Self::InvalidUtf8 => f.pad("file is not valid utf-8"),
             Self::Package(error) => error.fmt(f),
-            Self::Other => f.pad("failed to load file"),
+            Self::Other(Some(err)) => write!(f, "failed to load file ({err})"),
+            Self::Other(None) => f.pad("failed to load file"),
         }
     }
 }
@@ -374,31 +421,33 @@ impl From<FromUtf8Error> for FileError {
 }
 
 impl From<PackageError> for FileError {
-    fn from(error: PackageError) -> Self {
-        Self::Package(error)
+    fn from(err: PackageError) -> Self {
+        Self::Package(err)
     }
 }
 
 impl From<FileError> for EcoString {
-    fn from(error: FileError) -> Self {
-        eco_format!("{error}")
+    fn from(err: FileError) -> Self {
+        eco_format!("{err}")
     }
 }
 
 /// A result type with a package-related error.
 pub type PackageResult<T> = Result<T, PackageError>;
 
-/// An error that occured while trying to load a package.
+/// An error that occurred while trying to load a package.
+///
+/// Some variants have an optional string can give more details, if available.
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
 pub enum PackageError {
     /// The specified package does not exist.
     NotFound(PackageSpec),
     /// Failed to retrieve the package through the network.
-    NetworkFailed,
+    NetworkFailed(Option<EcoString>),
     /// The package archive was malformed.
-    MalformedArchive,
+    MalformedArchive(Option<EcoString>),
     /// Another error.
-    Other,
+    Other(Option<EcoString>),
 }
 
 impl std::error::Error for PackageError {}
@@ -409,44 +458,53 @@ impl Display for PackageError {
             Self::NotFound(spec) => {
                 write!(f, "package not found (searched for {spec})",)
             }
-            Self::NetworkFailed => f.pad("failed to load package (network failed)"),
-            Self::MalformedArchive => f.pad("failed to load package (archive malformed)"),
-            Self::Other => f.pad("failed to load package"),
+            Self::NetworkFailed(Some(err)) => {
+                write!(f, "failed to download package ({err})")
+            }
+            Self::NetworkFailed(None) => f.pad("failed to download package"),
+            Self::MalformedArchive(Some(err)) => {
+                write!(f, "failed to decompress package ({err})")
+            }
+            Self::MalformedArchive(None) => {
+                f.pad("failed to decompress package (archive malformed)")
+            }
+            Self::Other(Some(err)) => write!(f, "failed to load package ({err})"),
+            Self::Other(None) => f.pad("failed to load package"),
         }
     }
 }
 
 impl From<PackageError> for EcoString {
-    fn from(error: PackageError) -> Self {
-        eco_format!("{error}")
+    fn from(err: PackageError) -> Self {
+        eco_format!("{err}")
     }
 }
 
 /// Format a user-facing error message for an XML-like file format.
 pub fn format_xml_like_error(format: &str, error: roxmltree::Error) -> EcoString {
     match error {
-        roxmltree::Error::UnexpectedCloseTag { expected, actual, pos } => {
+        roxmltree::Error::UnexpectedCloseTag(expected, actual, pos) => {
             eco_format!(
-                "failed to parse {format}: found closing tag '{actual}' \
-                 instead of '{expected}' in line {}",
+                "failed to parse {format} (found closing tag '{actual}' \
+                 instead of '{expected}' in line {})",
                 pos.row
             )
         }
         roxmltree::Error::UnknownEntityReference(entity, pos) => {
             eco_format!(
-                "failed to parse {format}: unknown entity '{entity}' in line {}",
+                "failed to parse {format} (unknown entity '{entity}' in line {})",
                 pos.row
             )
         }
         roxmltree::Error::DuplicatedAttribute(attr, pos) => {
             eco_format!(
-                "failed to parse {format}: duplicate attribute '{attr}' in line {}",
+                "failed to parse {format} (duplicate attribute '{attr}' in line {})",
                 pos.row
             )
         }
         roxmltree::Error::NoRootNode => {
-            eco_format!("failed to parse {format}: missing root node")
+            eco_format!("failed to parse {format} (missing root node)")
         }
-        _ => eco_format!("failed to parse {format}"),
+        err => eco_format!("failed to parse {format} ({err})"),
     }
 }
