@@ -2,9 +2,25 @@
 
 pub mod fat;
 
-use std::fmt::{self, Debug, Formatter};
+#[macro_use]
+mod macros;
+mod bitset;
+mod deferred;
+mod hash;
+mod pico;
+mod scalar;
+
+pub use self::bitset::BitSet;
+pub use self::deferred::Deferred;
+pub use self::hash::LazyHash;
+pub use self::pico::PicoStr;
+pub use self::scalar::Scalar;
+
+use std::fmt::{Debug, Formatter};
 use std::hash::Hash;
+use std::iter::{Chain, Flatten, Rev};
 use std::num::NonZeroUsize;
+use std::ops::{Add, Deref, Div, Mul, Neg, Sub};
 use std::sync::Arc;
 
 use siphasher::sip128::{Hasher128, SipHasher13};
@@ -12,15 +28,15 @@ use siphasher::sip128::{Hasher128, SipHasher13};
 /// Turn a closure into a struct implementing [`Debug`].
 pub fn debug<F>(f: F) -> impl Debug
 where
-    F: Fn(&mut Formatter) -> fmt::Result,
+    F: Fn(&mut Formatter) -> std::fmt::Result,
 {
     struct Wrapper<F>(F);
 
     impl<F> Debug for Wrapper<F>
     where
-        F: Fn(&mut Formatter) -> fmt::Result,
+        F: Fn(&mut Formatter) -> std::fmt::Result,
     {
-        fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
             self.0(f)
         }
     }
@@ -103,83 +119,32 @@ where
     }
 }
 
-/// Format pieces separated with commas and a final "and" or "or".
-pub fn separated_list(pieces: &[impl AsRef<str>], last: &str) -> String {
-    let mut buf = String::new();
-    for (i, part) in pieces.iter().enumerate() {
-        match i {
-            0 => {}
-            1 if pieces.len() == 2 => {
-                buf.push(' ');
-                buf.push_str(last);
-                buf.push(' ');
-            }
-            i if i + 1 == pieces.len() => {
-                buf.push_str(", ");
-                buf.push_str(last);
-                buf.push(' ');
-            }
-            _ => buf.push_str(", "),
-        }
-        buf.push_str(part.as_ref());
-    }
-    buf
+/// Adapter for reversing iterators conditionally.
+pub trait MaybeReverseIter {
+    type RevIfIter;
+
+    /// Reverse this iterator (apply .rev()) based on some condition.
+    fn rev_if(self, condition: bool) -> Self::RevIfIter
+    where
+        Self: Sized;
 }
 
-/// Format a comma-separated list.
-///
-/// Tries to format horizontally, but falls back to vertical formatting if the
-/// pieces are too long.
-pub fn pretty_comma_list(pieces: &[impl AsRef<str>], trailing_comma: bool) -> String {
-    const MAX_WIDTH: usize = 50;
+impl<I: Iterator + DoubleEndedIterator> MaybeReverseIter for I {
+    type RevIfIter =
+        Chain<Flatten<std::option::IntoIter<I>>, Flatten<std::option::IntoIter<Rev<I>>>>;
 
-    let mut buf = String::new();
-    let len = pieces.iter().map(|s| s.as_ref().len()).sum::<usize>()
-        + 2 * pieces.len().saturating_sub(1);
+    fn rev_if(self, condition: bool) -> Self::RevIfIter
+    where
+        Self: Sized,
+    {
+        let (maybe_self_iter, maybe_rev_iter) =
+            if condition { (None, Some(self.rev())) } else { (Some(self), None) };
 
-    if len <= MAX_WIDTH {
-        for (i, piece) in pieces.iter().enumerate() {
-            if i > 0 {
-                buf.push_str(", ");
-            }
-            buf.push_str(piece.as_ref());
-        }
-        if trailing_comma {
-            buf.push(',');
-        }
-    } else {
-        for piece in pieces {
-            buf.push_str(piece.as_ref().trim());
-            buf.push_str(",\n");
-        }
+        maybe_self_iter
+            .into_iter()
+            .flatten()
+            .chain(maybe_rev_iter.into_iter().flatten())
     }
-
-    buf
-}
-
-/// Format an array-like construct.
-///
-/// Tries to format horizontally, but falls back to vertical formatting if the
-/// pieces are too long.
-pub fn pretty_array_like(parts: &[impl AsRef<str>], trailing_comma: bool) -> String {
-    let list = pretty_comma_list(parts, trailing_comma);
-    let mut buf = String::new();
-    buf.push('(');
-    if list.contains('\n') {
-        buf.push('\n');
-        for (i, line) in list.lines().enumerate() {
-            if i > 0 {
-                buf.push('\n');
-            }
-            buf.push_str("  ");
-            buf.push_str(line);
-        }
-        buf.push('\n');
-    } else {
-        buf.push_str(&list);
-    }
-    buf.push(')');
-    buf
 }
 
 /// Check if the [`Option`]-wrapped L is same to R.
@@ -187,5 +152,94 @@ pub fn option_eq<L, R>(left: Option<L>, other: R) -> bool
 where
     L: PartialEq<R>,
 {
-    left.map(|v| v == other).unwrap_or(false)
+    left.is_some_and(|v| v == other)
+}
+
+/// A container around a static reference that is cheap to clone and hash.
+#[derive(Debug)]
+pub struct Static<T: 'static>(pub &'static T);
+
+impl<T> Deref for Static<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        self.0
+    }
+}
+
+impl<T> Copy for Static<T> {}
+
+impl<T> Clone for Static<T> {
+    fn clone(&self) -> Self {
+        *self
+    }
+}
+
+impl<T> Eq for Static<T> {}
+
+impl<T> PartialEq for Static<T> {
+    fn eq(&self, other: &Self) -> bool {
+        std::ptr::eq(self.0, other.0)
+    }
+}
+
+impl<T> Hash for Static<T> {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        state.write_usize(self.0 as *const _ as _);
+    }
+}
+
+/// Generic access to a structure's components.
+pub trait Get<Index> {
+    /// The structure's component type.
+    type Component;
+
+    /// Borrow the component for the specified index.
+    fn get_ref(&self, index: Index) -> &Self::Component;
+
+    /// Borrow the component for the specified index mutably.
+    fn get_mut(&mut self, index: Index) -> &mut Self::Component;
+
+    /// Convenience method for getting a copy of a component.
+    fn get(self, index: Index) -> Self::Component
+    where
+        Self: Sized,
+        Self::Component: Copy,
+    {
+        *self.get_ref(index)
+    }
+
+    /// Convenience method for setting a component.
+    fn set(&mut self, index: Index, component: Self::Component) {
+        *self.get_mut(index) = component;
+    }
+}
+
+/// A numeric type.
+pub trait Numeric:
+    Sized
+    + Debug
+    + Copy
+    + PartialEq
+    + Neg<Output = Self>
+    + Add<Output = Self>
+    + Sub<Output = Self>
+    + Mul<f64, Output = Self>
+    + Div<f64, Output = Self>
+{
+    /// The identity element for addition.
+    fn zero() -> Self;
+
+    /// Whether `self` is zero.
+    fn is_zero(self) -> bool {
+        self == Self::zero()
+    }
+
+    /// Whether `self` consists only of finite parts.
+    fn is_finite(self) -> bool;
+}
+
+/// Round a float to two decimal places.
+pub fn round_2(value: f64) -> f64 {
+    (value * 100.0).round() / 100.0
 }
